refactor: extract PostTranscriptionStatusPresenter.swift from RecordingStatusMessageProvider.swift
Splits `PostTranscriptionStatusPresenter` (distinct concern from
recording status: coordinates post-transcription status/window
messages) out of RecordingStatusMessageProvider.swift into its own
file. RecordingStatusMessageProvider.swift retains
RecordingStatusMessageSnapshot + RecordingStatusMessageProvider.

Byte-identical move of lines 65-100. Both files import Combine +
Foundation.

RecordingStatusMessageProvider.swift: 100 → 63 lines. Tests: 667.
